(PE-32039) HSBC SSL Escalation
This commit is a temporary troubleshooting change.

The changes here are gathering data to understand why the splunk module
is not able to validate the certificates protecting the HEC end point
on the HSBC splunk server.

The goal is to manually retrieve the end point certificate, and perform
validation against the ca certificate provided to the module, and write
the result of the manual validation to a file.

It will then write a string representation of the certificates to files
to ensure that the ruby gems in use by Puppet can properly parse them
and to understand how those gems perceive the certificate chain for
each of the certificates.

This change will write three files.

1. /tmp/cert-validated
2. /tmp/ca_cert
3. /tmp/server_cert

If the validation code encounters any errors it will write a fourth file
called `/tmp/ssl-troubleshooting-error-log`.

All four of these files should be checked for content.